requirements.txt
pocketflow>=0.0.1
httpx>=0.27.0
rich>=13.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
ddgs>=8.0.0
pydantic>=2.0
